+++ src/main/java/ch/ethz/matsim/discrete_mode_choice/model/constraints/AbstractTourConstraint.java
@@ -2,6 +2,7 @@ package ch.ethz.matsim.discrete_mode_choice.model.constraints;
 
 import java.util.List;
 
+import ch.ethz.matsim.discrete_mode_choice.model.DiscreteModeChoiceTrip;
 import ch.ethz.matsim.discrete_mode_choice.model.tour_based.TourCandidate;
 import ch.ethz.matsim.discrete_mode_choice.model.tour_based.TourConstraint;
 
@@ -13,12 +14,14 @@ import ch.ethz.matsim.discrete_mode_choice.model.tour_based.TourConstraint;
  */
 public abstract class AbstractTourConstraint implements TourConstraint {
 	@Override
-	public boolean validateBeforeEstimation(List<String> modes, List<List<String>> previousModes) {
+	public boolean validateBeforeEstimation(List<DiscreteModeChoiceTrip> tour, List<String> modes,
+			List<List<String>> previousModes) {
 		return true;
 	}
 
 	@Override
-	public boolean validateAfterEstimation(TourCandidate candidate, List<TourCandidate> previousCandidates) {
+	public boolean validateAfterEstimation(List<DiscreteModeChoiceTrip> tour, TourCandidate candidate,
+			List<TourCandidate> previousCandidates) {
 		return true;
 	}
 }