+++ src/main/java/ch/ethz/matsim/discrete_mode_choice/modules/DiscreteModeChoiceConfigurator.java
@@ -63,9 +63,7 @@ public final class DiscreteModeChoiceConfigurator {
 		dmcConfig.setTourEstimator(EstimatorModule.UNIFORM);
 		dmcConfig.setTourFinder(TourFinderModule.PLAN_BASED);
 
-		dmcConfig.getVehicleTourConstraintConfig().setRequireContinuity(Arrays.asList(smcConfig.getChainBasedModes()));
-		dmcConfig.getVehicleTourConstraintConfig().setRequireStartAtHome(Arrays.asList(smcConfig.getChainBasedModes()));
-		dmcConfig.getVehicleTourConstraintConfig().setRequireEndAtHome(Arrays.asList(smcConfig.getChainBasedModes()));
+		dmcConfig.getVehicleTourConstraintConfig().setRestrictedModes(Arrays.asList(smcConfig.getChainBasedModes()));
 
 		if (smcConfig.considerCarAvailability()) {
 			dmcConfig.setModeAvailability(ModeAvailabilityModule.CAR);