+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+- Fix buggy vehicle constraints
+- Make inference of origin/destination facility in AbstractTripRouterEstimator compatible with PlanRouter
+- Fix MNL selection
+
 **1.0.2**
 
 - Fix 'restricted mode' setters for LinkAttributeConstraint and ShapeFileConstraint configuration